+++ R/lbt05.R
@@ -4,7 +4,7 @@
 #'
 #' @inheritParams gen_args
 #' @param map (`data.frame`) with two columns mapping of parameters code (`PARAMCD`) to directions of abnormality
-#' (`ABN_DIR`). If `NUll`, all combinations are used.
+#' (`ABN_DIR`). If a parameter is not in the `map` or if `map` is `NULL`, both directions are analyzed.
 #' @returns the main function returns an `rtables` object.
 #'
 #' @details
@@ -21,7 +21,7 @@
 lbt05_main <- function(adam_db,
                        arm_var = "ACTARM",
                        lbl_overall = NULL,
-                       map = NULL,
+                       map = lab_paramcd_abn_dir(),
                        ...) {
   assert_all_tablenames(adam_db, c("adsl", "adlb"))
   assert_string(arm_var)
@@ -38,31 +38,41 @@ lbt05_main <- function(adam_db,
   lbl_anrind <- var_labels_for(adam_db$adlb, "ABN_DIR")
   lbl_param <- var_labels_for(adam_db$adlb, "PARAM")
 
-  map <- if (is.null(map)) {
-    expand.grid(
-      PARAM = levels(adam_db$adlb$PARAM),
-      ABN_DIR = c("Low", "High"),
-      stringsAsFactors = FALSE
-    ) %>%
-      arrange(.data$PARAM, desc(.data$ABN_DIR))
-  } else {
-    matching <- adam_db$adlb %>%
-      dplyr::select(.data$PARAMCD, .data$PARAM) %>%
-      dplyr::distinct()
 
-    map %>%
-      dplyr::filter(.data$PARAMCD %in% matching$PARAMCD) %>%
-      dplyr::left_join(matching, by = "PARAMCD") %>%
-      dplyr::select(.data$PARAM, .data$ABN_DIR) %>%
-      mutate(across(everything(), as.character))
+  # This is the only way to get the correspondence between PARAM and PARAMCD if some levels are missing.
+  matching <- data.frame(
+    PARAMCD = levels(adam_db$adlb$PARAMCD),
+    PARAM = levels(adam_db$adlb$PARAM)
+  )
+
+  un_mapped <- adam_db$adlb %>%
+    dplyr::select(PARAM, PARAMCD) %>%
+    distinct() %>%
+    tidyr::drop_na() %>%
+    dplyr::anti_join(matching, by = c("PARAM", "PARAMCD"))
+
+  if (nrow(un_mapped) > 0) {
+    rlang::abort("unanticipated mapping! Please check for unique correspondence of PARAM and PARAMCD")
   }
+
+  map <- map %||% expand.grid(ABN_DIR = c("Low", "High"), PARAMCD = levels(adam_db$adlb$PARAM))
+
+  # Add both directions by default.
+  map_paramcd <- matching %>%
+    left_join(tidyr::nest(map, ABN_DIR = ABN_DIR), by = "PARAMCD") %>%
+    rowwise() %>%
+    mutate(ABN_DIR = ifelse((is.null(.data$ABN_DIR)), list(c("Low", "High")), ABN_DIR)) %>%
+    tidyr::unnest(ABN_DIR) %>%
+    dplyr::select(.data$PARAM, .data$ABN_DIR) %>%
+    mutate(across(everything(), as.character))
+
 
   lyt <- lbt05_lyt(
     arm_var = arm_var,
     lbl_overall = lbl_overall,
     lbl_param = lbl_param,
     lbl_anrind = lbl_anrind,
-    map = map
+    map = map_paramcd
   )
 
   tbl <- build_table(lyt, adam_db$adlb, alt_counts_df = adam_db$adsl)
@@ -162,3 +172,80 @@ lbt05 <- chevron_t(
   preprocess = lbt05_pre,
   postprocess = lbt05_post
 )
+
+
+
+#' @describeIn lbt05
+#'
+#' @returns a `data.frame` with the direction of abnormality of each lab parameter code.
+#' @export
+#' @examples
+#' # example code
+#' head(lab_paramcd_abn_dir())
+#'
+lab_paramcd_abn_dir <- function() {
+  no_suffix <- c(
+    HCRIT = c("Low", "High"),
+    HGB = c("Low", "High"),
+    WBC = c("Low", "High"),
+    PLATE = c("Low", "High"),
+    MCH = c("Low", "High"),
+    MCHC = c("Low", "High"),
+    MCV = c("Low", "High"),
+    RBC = c("Low", "High"),
+    BANDS = c("High"),
+    BANDSF = c("High"),
+    BASOS = c("High"),
+    BASOSF = c("High"),
+    LYMPH = c("Low", "High"),
+    LYMPHF = c("Low", "High"),
+    MONOS = c("High"),
+    MONOSF = c("High"),
+    NEUTR = c("Low", "High"),
+    NEUTRF = c("Low", "High"),
+    EOSIN = c("High"),
+    EOSINF = c("High"),
+    PTINR = c("High"),
+    APTT = c("High"),
+    FIB = c("Low"),
+    AST = c("High"),
+    LDH = c("High"),
+    CPK = c("High"),
+    CPKMB = c("High"),
+    ALKPH = c("High"),
+    ALT = c("High"),
+    TBILI = c("High"),
+    DBILI = c("High"),
+    GGT = c("High"),
+    BUN = c("High"),
+    CREATN = c("High"),
+    T3 = c("Low", "High"),
+    T4 = c("Low", "High"),
+    T4FREE = c("Low", "High"),
+    TSH = c("High"),
+    ALBUM = c("Low"),
+    TPROT = c("Low", "High"),
+    TRIG = c("High"),
+    CHOLES = c("High"),
+    LDL = c("High"),
+    HDL = c("Low"),
+    CHLOR = c("Low", "High"),
+    POTAS = c("Low", "High"),
+    SODIUM = c("Low", "High"),
+    BICARB = c("Low", "High"),
+    CALCUM = c("Low", "High"),
+    PHOSAT = c("Low", "High"),
+    FASTGL = c("Low", "High"),
+    URACID = c("High")
+  )
+
+  # add suffixes to list name for each lab parameter code
+  res <- c()
+  for (i in c("SI", "CV", "LS")) {
+    res <- c(res, setNames(no_suffix, paste0(names(no_suffix), i)))
+  }
+
+  res %>%
+    tibble::enframe(name = "PARAMCD", value = "ABN_DIR") %>%
+    as.data.frame()
+}